no message
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@64 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/CStr.cpp
@@ -219,7 +219,7 @@ void CStr::Replace(const CStr &ToReplace, const CStr &ReplaceWith)
 // returns a trimed string, removes whitespace from the left/right/both
 CStr CStr::Trim(PS_TRIM_MODE Mode)
 {
-	size_t Left, Right;
+	int Left, Right;
 	
 	
 	switch (Mode)
